move getchanneldata, gethighestrank in separate lib
