+++ lib/getChannelData.tsx
@@ -1,0 +1,5 @@
+import { CHANNELS } from './consts';
+
+export function getChannelData(channelId: any) {
+  return CHANNELS.find((channel) => channel.value === channelId);
+}
