+++ lib/getHighestRank.tsx
@@ -1,0 +1,3 @@
+export function getHighestRank(validRanks: any[]) {
+  return validRanks.length === 0 ? 0 : Math.min(...validRanks);
+}
